pyproject.toml
[tool.poetry]
name = "codeswarm"
version = "0.1.0"
description = "Self-improving multi-agent AI coding system"
authors = ["Blake Ledden"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
langgraph = "^0.0.45"
neo4j = "^5.15.0"
openai = "^1.6.0"
aiohttp = "^3.9.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
galileo-observe = "^0.3.0"
browser-use = "^0.2.0"
workos = "^4.3.0"
weave = "^0.50.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
ruff = "^0.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
